rag_chatbot/rag_chatbot/requirements.txt
# Core RAG dependencies
chromadb>=0.4.22
sentence-transformers>=2.7.0
langchain>=0.1.0
langchain-community>=0.0.20

# Document loading
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
unstructured>=0.12.0

# LLM providers
anthropic>=0.25.0
openai>=1.14.0

# Evaluation
ragas>=0.1.7
datasets>=2.18.0

# UI
gradio>=4.20.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.6.0
nltk>=3.8.1
tqdm>=4.66.0
pandas>=2.2.0
numpy>=1.26.0

# Architecture study
transformers>=4.38.0
torch>=2.2.0
